Move downloadFileToDestination to common; Replace downloadCpptoolsJsonAsync
