+++ Extension/src/abTesting.ts
@@ -83,45 +83,11 @@ export class ABTestSettings {
         }
     }
 
-    // NOTE: Code is copied from DownloadPackage in packageManager.ts, but with ~75% fewer lines.
-    private downloadCpptoolsJsonAsync(urlString): Promise<void> {
-        return new Promise<void>((resolve, reject) => {
-            let parsedUrl: url.Url = url.parse(urlString);
-            let request: ClientRequest = https.request({
-                host: parsedUrl.host,
-                path: parsedUrl.path,
-                agent: util.getHttpsProxyAgent(),
-                rejectUnauthorized: vscode.workspace.getConfiguration().get("http.proxyStrictSSL", true)
-            }, (response) => {
-                if (response.statusCode === 301 || response.statusCode === 302) {
-                    let redirectUrl: string | string[];
-                    if (typeof response.headers.location === "string") {
-                        redirectUrl = response.headers.location;
-                    } else {
-                        redirectUrl = response.headers.location[0];
-                    }
-                    return resolve(this.downloadCpptoolsJsonAsync(redirectUrl)); // Redirect - download from new location
-                }
-                if (response.statusCode !== 200) {
-                    return reject();
-                }
-                let downloadedBytes = 0; // tslint:disable-line
-                let cppToolsJsonFile: fs.WriteStream = fs.createWriteStream(util.getExtensionFilePath(localConfigFile));
-                response.on('data', (data) => { downloadedBytes += data.length; });
-                response.on('end', () => { cppToolsJsonFile.close(); });
-                cppToolsJsonFile.on('close', () => { resolve(); this.updateSettingsAsync(); });
-                response.on('error', (error) => { reject(); });
-                response.pipe(cppToolsJsonFile, { end: false });
-            });
-            request.on('error', (error) => { reject(); });
-            request.end();
-        });
-    }
-
     private downloadCpptoolsJsonPkgAsync(): Promise<void> {
         let hasError: boolean = false;
         let telemetryProperties: { [key: string]: string } = {};
-        return this.downloadCpptoolsJsonAsync("https://go.microsoft.com/fwlink/?linkid=852750")
+        const localConfigPath: string = util.getExtensionFilePath(localConfigFile);
+        return util.downloadFileToDestination("https://go.microsoft.com/fwlink/?linkid=852750", localConfigPath)
             .catch((error) => {
                 // More specific error info is not likely to be helpful, and we get detailed download data from the initial install.
                 hasError = true;
